requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
PyMySQL==1.1.0
cryptography==41.0.7